requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
google-generativeai==0.3.0
xmltodict==0.13.0 